server/requirements.txt
fastapi
fewsats
pydantic
python-dotenv
uvicorn
